Make some items that don't need to be size_t guint.
Those sizes are limited by the packet sizes we support, and we only
support a maximum packet size of 2^32.

This squelches some compiler warnings.

Remove some casts that this renders unnecessary.

Reviewed-on: https://code.wireshark.org/review/19279
Petri-Dish: Guy Harris <guy@alum.mit.edu>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>

+++ ui/gtk/rtp_analysis.c
@@ -783,7 +783,7 @@ rtp_packet_save_payload(tap_rtp_save_info_t    *saveinfo,
 			saveinfo->error_type = TAP_RTP_FILE_WRITE_ERROR;
 			return 0;
 		}
-		saveinfo->count += ((int)rtpinfo->info_payload_len - rtpinfo->info_padding_count);
+		saveinfo->count += rtpinfo->info_payload_len - rtpinfo->info_padding_count;
 
 		fflush(saveinfo->fp);
 		saveinfo->saved = TRUE;
